% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxycode.R
\name{d3.aspp}
\alias{d3.aspp}
\title{d3.aspp}
\usage{
d3.aspp(
  data,
  data1 = NULL,
  type_plot = "3D_plot",
  window,
  sigma = 15,
  theta = 40,
  phi = 35,
  col = "lightblue",
  title = " "
)
}
\arguments{
\item{data}{vector of x and y coordinates of data points.}

\item{data1}{vector of x and y coordinates of data points.}

\item{type_plot}{define the type of the plot either "3D_plot" or "3D_plot_interactive".}

\item{window}{window of observation, an object of class "owin".}

\item{sigma}{parameter for density function.}

\item{theta}{angles defining the viewing direction. theta gives the azimuthal direction and phi.}

\item{phi}{the colatitude.}

\item{col}{the color(s) of the surface facets. Transparent colors are ignored. This is recycled to the (nx-1)(ny-1) facets.}

\item{title}{character vector representing the title of the plot.}
}
\description{
This function for plotting of 3D kernel density visualization spatial point patterns.
}
\author{
Viktorian Miok
}
