% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxycode.R
\name{diff.aspp}
\alias{diff.aspp}
\title{diff.aspp}
\usage{
\method{diff}{aspp}(
  null,
  alternative,
  method = "friedman",
  nr_sq = 10,
  window,
  title = "",
  ribbon = TRUE,
  cex.axis = 1,
  cex.lab = 1,
  col = c("blue", "white", "red"),
  testonly = FALSE
)
}
\arguments{
\item{null}{the null model object to plot.}

\item{alternative}{the alternative model object to plot.}

\item{method}{method employed for non-parametric testing, "friedman" or "wilcoxon".}

\item{nr_sq}{number of squares for the polygon horizontal axis.}

\item{window}{window of observation, an object of class "owin".}

\item{title}{character indicate title of the plot.}

\item{ribbon}{logical flag indicating whether to display a ribbon showing the color map. Default is TRUE for new plots and FALSE for added plots.}

\item{cex.axis}{specify the size of the tick label numbers/text with a numeric value of length 1.}

\item{cex.lab}{specify the size of the axis label text with a numeric value of length 1.}

\item{col}{the colors to be used for the background.}

\item{testonly}{logical flag indicating whether to test or not.}
}
\description{
This function for testing the overall and local difference in spatial point patterns.
}
\author{
Viktorian Miok
}
