% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxycode.R
\name{distance.euclidean}
\alias{distance.euclidean}
\title{distance.euclidean}
\usage{
distance.euclidean(a, b)
}
\arguments{
\item{a}{Matrix of first set of locations where each row gives the coordinates of a particular point.}

\item{b}{Matrix of second set of locations where each row gives the coordinates of a particular point. If this is missing a is used.}
}
\description{
This is a function for calculating euclidean distance.
}
\author{
Viktorian Miok
}
