% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxycode.R
\name{get.data}
\alias{get.data}
\title{get.data}
\usage{
get.data(data, marker, diet, mouse, arc, cal_arc)
}
\arguments{
\item{data}{data frame. Data set of spatial point pattern.}

\item{marker}{character. Indicate which marker to include in data set.}

\item{diet}{numeric. Indicate  which diet to include in data set.}

\item{mouse}{numeric. Indicate which mouse to include in data set.}

\item{arc}{numeric. Indicate which arc to include in data set.}

\item{cal_arc}{logical flag indicating whether to average over the ARCs.}
}
\description{
This function which from whole data set extract data elements of interest.
}
\author{
Viktorian Miok
}
