% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxycode.R
\name{get.knn}
\alias{get.knn}
\title{get.knn}
\usage{
get.knn(
  train_labels,
  train_data,
  to_classify_data,
  k,
  distance = distance.euclidean,
  standardize = FALSE
)
}
\arguments{
\item{train_labels}{a character.}

\item{train_data}{a character.}

\item{to_classify_data}{a character.}

\item{k}{the maximum number of nearest neighbors to search. The default value is set to 10.}

\item{distance}{function for calculating the distance between the cells.}

\item{standardize}{logical whether or not to standardize.}
}
\description{
Fast k-nearest neighbor searching algorithms including a kd-tree, cover-tree and the algorithm implemented in class package.
}
\author{
Viktorian Miok
}
