% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxycode.R
\name{gg.aspp}
\alias{gg.aspp}
\title{gg.aspp}
\usage{
gg.aspp(
  data,
  type_plot = "scatter_plot",
  title = " ",
  legend.text = 15,
  legend.title = 15,
  axis.text = 15,
  axis.title = 15,
  plot.title = 15,
  xlim = c(0, 500),
  ylim = c(0, 500)
)
}
\arguments{
\item{data}{a vector of x and y coordinates of data point.}

\item{type_plot}{a character that indicate the type of the plot. On of the following: "hexagonal_heatmap", "scatter_plot", "polygon", "density_contour".}

\item{legend.text}{a character or expression vector of length ≥ 1 to appear in the legend.}

\item{legend.title}{a character string or length-one expression giving a title to be placed at the top of the legend.}

\item{axis.text}{a character for both x and y axis texts.}

\item{axis.title}{a character for both x and y axis title.}

\item{plot.title}{a character that represent the title of the plot.}

\item{xlim}{a observations not in this range will be dropped completely and not passed to any other layers. If a NA value is substituted for one of the limits that limit is automatically calculated.}

\item{ylim}{a Observations not in this range will be dropped completely and not passed to any other layers. If a NA value is substituted for one of the limits that limit is automatically calculated.}
}
\description{
This function for plotting for initial cellular spatial point pattern visualization.
}
\author{
Viktorian Miok
}
