% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxycode.R
\name{mdim.scal}
\alias{mdim.scal}
\title{mdim.scal}
\usage{
mdim.scal(dataX, dataY, dataG)
}
\arguments{
\item{dataX}{vector of elements that represent x-axis values of the point patterns.}

\item{dataY}{vector of elements that represent y-axis values of the point patterns.}

\item{dataG}{vector of elements that represent group values of the point patterns.}
}
\description{
This function for multidimensional scaling based on random forest proximity.
}
\author{
Viktorian Miok
}
