% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxycode.R
\name{neighbor.ratio}
\alias{neighbor.ratio}
\title{neighbor.ratio}
\usage{
neighbor.ratio(dataX, dataY, dataG, diam, per_marker = TRUE)
}
\arguments{
\item{dataX}{vector of elements that represent x-axis values of the point patterns.}

\item{dataY}{vector of elements that represent y-axis values of the point patterns.}

\item{dataG}{vector of elements that represent group values of the point patterns.}

\item{diam}{number represent the diameter.}

\item{per_marker}{a logical.}
}
\description{
This function calculate over all percentage of the cells in the neighborhood within the diameter.
}
\author{
Viktorian Miok
}
