% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxycode.R
\name{plot.2d.classify}
\alias{plot.2d.classify}
\title{plot.2d.classify}
\usage{
\method{plot}{`2d.classify`}(
  to_classify_labels,
  to_classify_data,
  classify_method,
  X1,
  X2,
  lengthX1 = 100,
  lengthX2 = 100,
  title = "",
  plot_class_rate = TRUE
)
}
\arguments{
\item{to_classify_labels}{a character.}

\item{to_classify_data}{a character.}

\item{classify_method}{a character}

\item{X1}{a character.}

\item{X2}{a character.}

\item{lengthX1}{a numeric indicating the number for x-axis.}

\item{lengthX2}{a numeric indicating the number for y-axis.}

\item{title}{a character indicating title of the plot.}

\item{plot_class_rate}{a logical whether to plot or not class rate.}
}
\description{
Function for a general classification method with two features to visualize.
}
\author{
Viktorian Miok
}
