% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxycode.R
\name{plot.2d.knn}
\alias{plot.2d.knn}
\title{plot.2d.knn}
\usage{
\method{plot}{`2d.knn`}(
  train_labels,
  train_data,
  k,
  X1,
  X2,
  distance = distance.euclidean,
  standardize = FALSE,
  to_classify_labels = train_labels,
  to_classify_data = train_data,
  lengthX1 = 100,
  lengthX2 = 100,
  plot_class_rate = TRUE
)
}
\arguments{
\item{train_labels}{a character.}

\item{train_data}{a character.}

\item{k}{the maximum number of nearest neighbors to search. The default value is set to 10.}

\item{X1}{a charter.}

\item{X2}{a charter.}

\item{distance}{function for calculating the distance between the cells.}

\item{standardize}{logical whether or not to standardize.}

\item{to_classify_labels}{a character.}

\item{to_classify_data}{a character.}

\item{lengthX1}{a numeric indicating the number for x-axis.}

\item{lengthX2}{a numeric indicating the number for y-axis.}

\item{plot_class_rate}{a logical whether to plot or not class rate.}
}
\description{
Function for simple k-NN with two features to visualize class boundaries.
}
\author{
Viktorian Miok
}
