% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxycode.R
\name{randF.class}
\alias{randF.class}
\title{randF.class}
\usage{
randF.class(dataX, dataY, dataG)
}
\arguments{
\item{dataX}{vector of elements that represent x-axis values of the point patterns.}

\item{dataY}{vector of elements that represent y-axis values of the point patterns.}

\item{dataG}{vector of elements that represent group values of the point patterns.}
}
\description{
This function for plotting of random forest classification.
}
\author{
Viktorian Miok
}
