% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxycode.R
\name{spp.viz}
\alias{spp.viz}
\title{spp.viz}
\usage{
spp.viz(
  data,
  data1 = NULL,
  test_null = NULL,
  test_alt = NULL,
  method = "friedman",
  nr_sq = 10,
  z_lim = NULL,
  type_plot = "statspat_plot",
  title = " ",
  window,
  legend.text = 15,
  legend.title = 15,
  axis.text = 15,
  axis.title = 15,
  plot.title = 15,
  cex.axis = 1,
  cex.lab = 1,
  col = c("blue", "white", "red"),
  xlim,
  ylim,
  num_size = 1.2,
  num_size_sig = 1.5,
  ribbon = FALSE,
  contour = TRUE,
  bias = NULL,
  add_star_sig = FALSE,
  test_overall = TRUE
)
}
\arguments{
\item{data}{Vector of x and y coordinates of data point.}

\item{data1}{Vector of x and y coordinates of data point.}

\item{test_null}{the null model object to plot.}

\item{test_alt}{the alternative model object to plot.}

\item{method}{method employed for testing "friedman" or "wilcoxon".}

\item{nr_sq}{number of squares for the polygon horizontal axis.}

\item{z_lim}{plotting limits.}

\item{type_plot}{type of the plot to be generated "hexagonal_heatmap", "scatter_plot", "polygon", "density_contour","statspat_square_plot", "statspat_plot_subtract", "statspat_plot"}

\item{title}{character indicate title of the plot.}

\item{window}{window of observation, an object of class "owin".}

\item{legend.text}{a character or expression vector of length ≥ 1 to appear in the legend.}

\item{legend.title}{a character string or length-one expression giving a title to be placed at the top of the legend.}

\item{axis.text}{a character for both x and y axis texts.}

\item{axis.title}{a character for both x and y axis title.}

\item{plot.title}{a character that represent the title of the plot.}

\item{cex.axis}{specify the size of the tick label numbers/text with a numeric value of length 1.}

\item{cex.lab}{specify the size of the axis label text with a numeric value of length 1.}

\item{col}{The colors to be used for the background.}

\item{xlim}{a observations not in this range will be dropped completely and not passed to any other layers. If a NA value is substituted for one of the limits that limit is automatically calculated.}

\item{ylim}{a observations not in this range will be dropped completely and not passed to any other layers. If a NA value is substituted for one of the limits that limit is automatically calculated.}

\item{num_size}{number indicating the size of the labels.}

\item{num_size_sig}{number indicating the size of the labels.}

\item{ribbon}{logical flag indicating whether to display a ribbon showing the colour map. Default is TRUE for new plots and FALSE for added plots.}

\item{contour}{create a contour plot, or add contour lines to an existing plot.}

\item{bias}{bias >1 puts more colors at high values.}

\item{add_star_sig}{logical flag indicating whether to add or not star in the square.}

\item{test_overall}{logical flag indicating whether to test or not.}
}
\description{
This function for visualization and testing of spatial point patterns - all combined.
}
\author{
Viktorian Miok
}
