% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxycode.R
\name{test.saptcorr}
\alias{test.saptcorr}
\title{test.saptcorr}
\usage{
test.saptcorr(dataX, dataY, dataG, method)
}
\arguments{
\item{dataX}{vector of elements that represent x-axis values of the point patterns.}

\item{dataY}{vector of elements that represent y-axis values of the point patterns.}

\item{dataG}{vector of elements that represent group values of the point patterns.}

\item{method}{character representing the statistical method employed, either Mentel or Moran.I}
}
\description{
This function for spatial auto-correlation employing Moran.I and Mantel test.
}
\author{
Viktorian Miok
}
