% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxycode.R
\name{test.squares}
\alias{test.squares}
\title{test.squares}
\usage{
test.squares(
  null = dat1,
  alt = dat2,
  method = "friedman",
  window = window,
  nr_sq = 10
)
}
\arguments{
\item{null}{the null model object to plot.}

\item{alt}{the alternative model object to plot.}

\item{method}{method employed for testing "friedman" or "wilcoxon".}

\item{window}{window of observation, an object of class "owin".}

\item{nr_sq}{number of squares for the poligon horizontal axis.}
}
\description{
This function for testing of square comparison of cellular spatial point patterns.
}
\author{
Viktorian Miok
}
